QuestionAPI/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.9.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==47.0.0
distro==1.9.0
dj-database-url==3.1.2
Django==5.2.6
django-contrib-postgres==0.0.1
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.49.2
google-genai==1.74.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
packaging==26.1
psycopg2-binary==2.9.10
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.11.0
python-dotenv==1.1.1
requests==2.33.1
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
websockets==16.0
whitenoise==6.12.0
